# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sunline APIs",
        "description": "API documentation for Sunline Energy.",
        "version": "2.0.0"
    },
    "paths": {
        "/api/v1/userLogin": {
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "User Login",
                "operationId": "e9ba8e14abe395ed3b01c9df50cd505b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "kap@yopmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "kap@yopmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "your_access_token"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 31536000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": []
            }
        },
        "/api/v1/createLead": {
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Create a new lead",
                "operationId": "c7104190cf8a1fc591af15a7a25ef7d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+911234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Street, City, Country"
                                    },
                                    "lead_source": {
                                        "type": "string",
                                        "example": "5"
                                    },
                                    "customer_type": {
                                        "type": "string",
                                        "example": "Residential"
                                    },
                                    "request_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-03"
                                    },
                                    "suburb": {
                                        "type": "string",
                                        "example": "Sydney"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "NSW"
                                    },
                                    "postcode": {
                                        "type": "string",
                                        "example": "2000"
                                    },
                                    "preferred_time_of_contact": {
                                        "type": "string",
                                        "example": "Morning"
                                    },
                                    "when_customer_wants_jobs_done": {
                                        "type": "string",
                                        "example": "Next Week"
                                    },
                                    "rejection_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/reject/lead123"
                                    },
                                    "power_bill_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/uploads/powerbill123.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lead created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Leads",
            "description": "Leads"
        }
    ]
}